模板水印架构重构，提取可复用内容
