+++ src/views/frame-watermark/constants/colors.js
@@ -1,0 +1,21 @@
+// 莫兰蒂色系推荐
+export const morandiColors = [
+  { name: '雾霾蓝', color: '#92A5CF' },
+  { name: '豆沙绿', color: '#B7C4CF' },
+  { name: '暖米灰', color: '#D6C5B9' },
+  { name: '玫瑰灰', color: '#C8A2C8' },
+  { name: '烟粉色', color: '#E6D2D2' },
+  { name: '薄荷绿', color: '#ACCFCC' },
+  { name: '奶茶色', color: '#D4C0A1' }
+]
+
+// 优雅渐变色组合推荐
+export const gradientCombos = [
+  { name: '日落黄昏', colors: ['#FFB347', '#FF8C94'] },
+  { name: '海洋蓝调', colors: ['#92A5CF', '#6FB3D2'] },
+  { name: '森林薄雾', colors: ['#ACCFCC', '#B7C4CF'] },
+  { name: '玫瑰晨曦', colors: ['#E6D2D2', '#C8A2C8'] },
+  { name: '奶茶拿铁', colors: ['#D4C0A1', '#D6C5B9'] },
+  { name: '梦幻紫调', colors: ['#C8A2C8', '#B19CD9'] },
+  { name: '温柔米色', colors: ['#F5E6D3', '#E8D5C4'] }
+]
