+++ src/views/frame-watermark/constants/fonts.js
@@ -1,0 +1,52 @@
+// 字体显示名称映射
+export const fontDisplayNames = {
+  'ZCOOL XiaoWei': '站酷小薇',
+  'Ma Shan Zheng': '马善政毛笔楷书',
+  'Liu Jian Mao Cao': '钟齐流江毛笔草'
+}
+
+// 完整字体库定义
+export const fontLibrary = {
+  signature: {
+    name: '签名艺术体',
+    fonts: [
+      'Dancing Script', 'Pacifico', 'Kaushan Script', 'Great Vibes', 'Allura'
+    ]
+  },
+  chinese: {
+    name: '中文书法体',
+    fonts: [
+      'ZCOOL XiaoWei', 'Ma Shan Zheng', 'Liu Jian Mao Cao'
+    ]
+  },
+  modern: {
+    name: '现代艺术体',
+    fonts: [
+      'Lobster', 'Fredoka One', 'Comfortaa', 'Righteous', 'Bungee'
+    ]
+  },
+  serif: {
+    name: '优雅衬线体',
+    fonts: [
+      'Playfair Display', 'Cormorant Garamond', 'Crimson Text', 'Lora', 'Merriweather'
+    ]
+  },
+  sansSerif: {
+    name: '现代简约体',
+    fonts: [
+      'Inter', 'Source Sans Pro', 'Open Sans'
+    ]
+  },
+  monospace: {
+    name: '专业等宽体',
+    fonts: [
+      'JetBrains Mono', 'Source Code Pro', 'Roboto Mono'
+    ]
+  },
+  system: {
+    name: '系统字体',
+    fonts: [
+      '微软雅黑', 'Arial', 'Times New Roman', 'Helvetica', 'Georgia', 'Verdana'
+    ]
+  }
+}
